requirements.txt
black==22.8.0
Django==4.1.1
djangorestframework==3.13.1
pytest==7.1.3
djangorestframework-simplejwt==5.2.0
Pillow
drf-yasg==1.21.3
